info über anzahl in weitere komponente ausgelagert
